+++ example/lib/var.dart
@@ -11,12 +11,12 @@ int get int1 => _int1.value;
 set int1(value) => _int1.value = value;
 
 /// Mediator Variable: int2
-final _int2 = 0.rx; // or `final _int2 = Rx(0);`
+final _int2 = 0.signal; // or `final _int2 = Signal(0);`
 int get int2 => _int2.value;
 set int2(value) => _int2.value = value;
 
 /// Mediator Variable: int3
-final _int3 = Rx(0); // or `final _int3 = 0.rx;`
+final _int3 = Signal(0); // or `final _int3 = 0.signal;`
 int get int3 => _int3.value;
 set int3(value) => _int3.value = value;
 
@@ -27,7 +27,7 @@ final _sum = Rx(() {
   if (res <= 10) {
     return res;
   }
-  return "excess upper bound";
+  return "excess upper bound($res)";
 });
 get sum => _sum.value;
 set sum(value) => _sum.value = value;
@@ -78,8 +78,10 @@ void updateListItem() {
   if (data.length >= maxItems) data.clear();
 
   //* Make an update to the collection type mediator variable.
-  // data.add(ListItem(itemName, units, color));
-  _data.notify.add(ListItem(itemName, units, color)); // notify  to rebuild
+  final item = ListItem(itemName, units, color);
+
+  // Notify to rebuild. Actually, `RxList.add` notifies itself.
+  _data.notify.add(item); // data.add(item);
 }
 
 /// Locale page