+++ lib/mediator/rx_map.dart
@@ -109,3 +109,11 @@ class RxMap<K, V> extends RxImpl<Map<K, V>> implements Map<K, V> {
 extension RxMapExt<K, V> on Map<K, V> {
   RxMap<K, V> get rx => RxMap<K, V>(this);
 }
+
+/// type alias
+typedef SignalMap<K, V> = RxMap<K, V>;
+
+/// `signal` extension of RxMap
+extension SignalMapExt<K, V> on Map<K, V> {
+  SignalMap<K, V> get signal => SignalMap<K, V>(this);
+}